+++ src/main/scala/com/github/simplesteph/ksm/source/FileSourceAcl.scala
@@ -1,8 +1,7 @@
 package com.github.simplesteph.ksm.source
 
-import java.io.{File, FileReader}
+import java.io.{File, FileReader, Reader}
 
-import com.github.simplesteph.ksm.parser.AclParser
 import com.typesafe.config.Config
 
 class FileSourceAcl extends SourceAcl {
@@ -27,14 +26,12 @@ class FileSourceAcl extends SourceAcl {
     * Uses a CSV parser on the file afterwards
     * @return
     */
-  override def refresh(aclParser: AclParser): Option[SourceAclResult] = {
+  override def refresh(): Option[Reader] = {
     val file = new File(filename)
     if (file.lastModified() > lastModified) {
       val reader = new FileReader(file)
-      val res = aclParser.aclsFromReader(reader)
-      reader.close()
       lastModified = file.lastModified()
-      Some(res)
+      Some(reader)
     } else {
       None
     }