+++ src/main/scala/com/github/simplesteph/ksm/source/GitHubSourceAcl.scala
@@ -1,11 +1,10 @@
 package com.github.simplesteph.ksm.source
 
-import java.io.StringReader
+import java.io.{Reader, StringReader}
 import java.nio.charset.Charset
 import java.util.Base64
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.github.simplesteph.ksm.parser.AclParser
 import com.typesafe.config.Config
 import org.slf4j.LoggerFactory
 import skinny.http.{HTTP, HTTPException, Request, Response}
@@ -48,10 +47,12 @@ class GitHubSourceAcl extends SourceAcl {
     tokenOpt = Try(config.getString(AUTH_TOKEN_CONFIG)).toOption
   }
 
-  override def refresh(aclParser: AclParser): Option[SourceAclResult] = {
+  override def refresh(): Option[Reader] = {
     val url =
       s"https://$hostname/repos/$user/$repo/contents/$filepath?ref=$branch"
     val request: Request = new Request(url)
+    // super important in order to properly fail in case a timeout happens for example
+    request.enableThrowingIOException(true)
 
     // authentication if present
     basicOpt.foreach(basic => {
@@ -72,21 +73,18 @@ class GitHubSourceAcl extends SourceAcl {
         val b64encodedContent =
           objectMapper.readTree(response.textBody).get("content").asText()
         val data = new String(
-          Base64.getDecoder.decode(
-            b64encodedContent.replace("\n", "").replace("\r", "")),
-          Charset.forName("UTF-8"))
+          Base64.getDecoder
+            .decode(b64encodedContent.replace("\n", "").replace("\r", "")),
+          Charset.forName("UTF-8")
+        )
         // use the CSV Parser
-        Some(aclParser.aclsFromReader(new StringReader(data)))
+        Some(new StringReader(data))
       case 304 =>
         None
       case _ =>
         // we got an http error so we propagate it
         log.warn(response.asString)
-        Some(
-          SourceAclResult(
-            Set(),
-            List(Try(
-              throw HTTPException(Some("Failure to fetch file"), response)))))
+        throw HTTPException(Some(response.asString), response)
     }
   }