+++ src/main/scala/com/github/simplesteph/ksm/source/SourceAcl.scala
@@ -1,6 +1,7 @@
 package com.github.simplesteph.ksm.source
 
-import com.github.simplesteph.ksm.parser.AclParser
+import java.io.Reader
+
 import com.typesafe.config.Config
 
 trait SourceAcl {
@@ -19,13 +20,13 @@ trait SourceAcl {
     * Refresh the current view on the external source of truth for Acl
     * Ideally this function is smart and does not pull the entire external Acl at every iteration
     * Return `None` if the Source Acls have not changed (usually using metadata).
-    * Return `Some(x)` if the Acls have changed. `x` represents the parsing and parsing errors if any
+    * Return `Some(reader)` if some text is being returned and should be parsed next
     * Note: the first call to this function should never return `None`.
     *
     * Kafka Security Manager will not update Acls in Kafka until there are no errors in the result
     * @return
     */
-  def refresh(aclParser: AclParser): Option[SourceAclResult]
+  def refresh(): Option[Reader]
 
   /**
     * Close all the necessary underlying objects or connections belonging to this instance