+++ src/test/scala/com/github/simplesteph/ksm/utils/ProtoConversionUtilsTest.scala
@@ -10,17 +10,23 @@ class ProtoConversionUtilsTest extends FlatSpec with Matchers {
 
   "resourceToPb" should "correctly convert a resource" in {
     val resource: Resource = Resource(Topic, "foobar", PatternType.LITERAL)
-    val resourcepb: ResourcePb = ResourcePb(name = "foobar", kafkaResourceType = ResourceTypePb.RESOURCE_TYPE_TOPIC, PatternTypePb.PATTERN_TYPE_LITERAL)
+    val resourcepb: ResourcePb = ResourcePb(
+      name = "foobar",
+      kafkaResourceType = ResourceTypePb.RESOURCE_TYPE_TOPIC,
+      PatternTypePb.PATTERN_TYPE_LITERAL
+    )
     ProtoConversionUtils.resourceToPb(resource) shouldBe resourcepb
   }
 
   "aclToPb" should "correctly convert a acl" in {
-    val acl: Acl = Acl(SecurityUtils.parseKafkaPrincipal("User:john"), Allow, "*", Read)
+    val acl: Acl =
+      Acl(SecurityUtils.parseKafkaPrincipal("User:john"), Allow, "*", Read)
     val aclpb: AclPb = AclPb(
       principal = Some(KafkaPrincipalPb("User", "john")),
       permissionType = PermissionTypePb.PERMISSION_TYPE_ALLOW,
       "*",
-      operationType = OperationTypePb.OPERATION_TYPE_READ)
+      operationType = OperationTypePb.OPERATION_TYPE_READ
+    )
     ProtoConversionUtils.aclToPb(acl) shouldBe aclpb
   }
 }