+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - TODO: Upgrade to Kafka 2.4.x (PR welcome)
 - Added Bitbucket Cloud as an ACL source
 - Added GitLab as an ACL souce
+- Massive refactor to better handle refresh and parsing exceptions
 
 ## [0.8 - 10/01/2020]
 - Added possibility to run AclSyncornizer just once and then stop KSM (Issue #56)